+++ examples/resources/routeros_interface_wireless_cap/resource.tf
@@ -1,0 +1,5 @@
+resource "routeros_interface_wireless_cap" "settings" {
+  discovery_interfaces = ["bridge1"]
+  enabled              = true
+  interfaces           = ["wlan1", "wlan2"]
+}