+++ routeros/resource_ip_cloud_advanced.go
@@ -1,0 +1,39 @@
+package routeros
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+/*
+{
+  "use-local-address":"false"
+}
+*/
+
+// https://help.mikrotik.com/docs/display/ROS/Cloud#Cloud-Advanced
+func ResourceIpCloudAdvanced() *schema.Resource {
+	resSchema := map[string]*schema.Schema{
+		MetaResourcePath: PropResourcePath("/ip/cloud/advanced"),
+		MetaId:           PropId(Id),
+
+		"use_local_address": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Default:     false,
+			Description: "An option whether to assign an internal router address to the dynamic DNS name.",
+		},
+	}
+
+	return &schema.Resource{
+		CreateContext: DefaultSystemCreate(resSchema),
+		ReadContext:   DefaultSystemRead(resSchema),
+		UpdateContext: DefaultSystemUpdate(resSchema),
+		DeleteContext: DefaultSystemDelete(resSchema),
+
+		Importer: &schema.ResourceImporter{
+			StateContext: schema.ImportStatePassthroughContext,
+		},
+
+		Schema: resSchema,
+	}
+}